shortTerm completed

+++ weather/src/main/java/com/practice/weather/shortTerm/expectation/repository/ShortTermExpectationRepository.java
@@ -3,6 +3,6 @@ package com.practice.weather.shortTerm.expectation.repository;
 import com.practice.weather.shortTerm.expectation.entity.ShortTermExpectationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShortTermExpectationRepository extends JpaRepository<ShortTermExpectationEntity, Integer>, ShortTermExpectationRepositoryCustom {
+public interface ShortTermExpectationRepository extends JpaRepository<ShortTermExpectationEntity, Long>, ShortTermExpectationRepositoryCustom {
 
 }
